backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiocache>=0.12.3",
    "fast-bitrix24>=1.8.5",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "tzdata>=2025.2",
    "uvicorn>=0.34.2",
]

[tool.setuptools]
py-modules = []
